Conflicts:
	BDT-JPA/src/META-INF/persistence.xml
	BDT-JPA/src/cl/minsal/divap/model/Seguimiento.java
	BDT-JPA/src/cl/minsal/divap/model/ServicioSalud.java
	BDT-Rest/src/minsal/divap/listener/EJBLookupListener.java
	BDT-ejb/ejbModule/minsal/divap/dao/DocumentDAO.java
	BDT-ejb/ejbModule/minsal/divap/dao/SeguimientoDAO.java
	BDT-ejb/ejbModule/minsal/divap/dao/ServicioSaludDAO.java
	BDT-ejb/ejbModule/minsal/divap/enums/TareasSeguimiento.java
	BDT-ejb/ejbModule/minsal/divap/enums/TipoDocumentosProcesos.java
	BDT-ejb/ejbModule/minsal/divap/service/CajaService.java
	BDT-ejb/ejbModule/minsal/divap/service/DocumentService.java
	BDT-ejb/ejbModule/minsal/divap/service/EstimacionFlujoCajaService.java
	BDT-ejb/ejbModule/minsal/divap/service/ProgramasService.java
	BDT-ejb/ejbModule/minsal/divap/service/SeguimientoService.java
	BDT-ejb/ejbModule/minsal/divap/service/ServicioSaludService.java
	BDT-web/src/main/java/cl/minsal/divap/controller/ProcesoOTRevisarConsolidacionController.java

merge cambios percapita con cambios Aldo y Leandro

+++ BDT-ejb/ejbModule/minsal/divap/vo/CumplimientoVO.java
@@ -1,8 +1,6 @@
 package minsal.divap.vo;
 
 import java.io.Serializable;
-import java.util.ArrayList;
-import java.util.List;
 
 public class CumplimientoVO implements Serializable{
 
@@ -13,9 +11,9 @@ public class CumplimientoVO implements Serializable{
 	private String desc_item1;
 	private String desc_item2;
 	private String desc_item3;
-	private Integer value_item1;
-	private Integer value_item2;
-	private Integer value_item3;
+	private Double value_item1;
+	private Double value_item2;
+	private Double value_item3;
 	public Integer getId_comuna() {
 		return id_comuna;
 	}
@@ -40,22 +38,22 @@ public class CumplimientoVO implements Serializable{
 	public void setDesc_item3(String desc_item3) {
 		this.desc_item3 = desc_item3;
 	}
-	public Integer getValue_item1() {
+	public Double getValue_item1() {
 		return value_item1;
 	}
-	public void setValue_item1(Integer value_item1) {
+	public void setValue_item1(Double value_item1) {
 		this.value_item1 = value_item1;
 	}
-	public Integer getValue_item2() {
+	public Double getValue_item2() {
 		return value_item2;
 	}
-	public void setValue_item2(Integer value_item2) {
+	public void setValue_item2(Double value_item2) {
 		this.value_item2 = value_item2;
 	}
-	public Integer getValue_item3() {
+	public Double getValue_item3() {
 		return value_item3;
 	}
-	public void setValue_item3(Integer value_item3) {
+	public void setValue_item3(Double value_item3) {
 		this.value_item3 = value_item3;
 	}
